requirements.txt
streamlit
requests
beautifulsoup4
readability-lxml
markdownify
sentence-transformers
chromadb
numpy
pandas
python-dotenv
loguru
